refactor: move the observer panel to @scenetest/checks/panel (0.13.0)
The observer panel is the viewing lens for checks' assertions —
dependency-free DOM code hooking window.__scenetest_report, usable under
any bundler with no Vite (the docs site does exactly this). It belongs
with the assertions, not the Vite plugin.

- @scenetest/checks gains the ./panel subpath (auto-initializing import)
- @scenetest/vite-plugin drops ./panels/observer (introduced in 0.12.0)
  and injects the panel from its checks dependency instead; the
  plugin-served niceties (click-to-editor, dashboard link) degrade
  gracefully without the plugin
- the recorder stays in vite-plugin: equally vite-free in principle, but
  it composes scene DSL and has no standalone consumer today
- docs drops its vite-plugin dependency (it only carried it for the
  panel) and its prebuild filter shrinks back to the checks chain

checks and vite-plugin bump to 0.13.0; the changelog notes that
@scenetest/checks-panel's npm deprecation message should be re-pointed
at @scenetest/checks/panel.
