+++ docs/package.json
@@ -6,7 +6,7 @@
   "sideEffects": false,
   "scripts": {
     "dev": "vite",
-    "build": "pnpm --filter \"@scenetest/vite-plugin...\" build && NITRO_PRESET=cloudflare_module vite build",
+    "build": "pnpm --filter \"@scenetest/checks...\" build && NITRO_PRESET=cloudflare_module vite build",
     "preview": "pnpm build && wrangler dev",
     "deploy": "pnpm build && wrangler deploy"
   },
@@ -21,8 +21,7 @@
     "react-markdown": "^10.1.0",
     "rehype-highlight": "^7.0.2",
     "rehype-raw": "^7.0.0",
-    "remark-gfm": "^4.0.1",
-    "@scenetest/vite-plugin": "workspace:*"
+    "remark-gfm": "^4.0.1"
   },
   "devDependencies": {
     "@types/hast": "^3.0.4",
