+++ packages/checks/src/panel/fs-viewer.ts
