+++ packages/checks/src/panel/fullscreen.ts
