+++ packages/checks/src/panel/render.ts
