+++ packages/checks/src/panel/styles.ts
