+++ packages/checks/src/panel/types.ts
